utils/jwt_helper.py
import jwt
from datetime import datetime, timedelta
from flask import request, jsonify
from functools import wraps
from config import Config

def generate_token(user_id):
    payload = {
        'user_id': user_id,
        'exp': datetime.utcnow() + timedelta(hours=1)
    }
    return jwt.encode(payload, Config.SECRET_KEY, algorithm='HS256')

def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None
        if 'Authorization' in request.headers:
            token = request.headers['Authorization'].split(" ")[1]

        if not token:
            return jsonify({'message': 'Token is missing!'}), 401

        try:
            data = jwt.decode(token, Config.SECRET_KEY, algorithms=['HS256'])
            user_id = data['user_id']
        except:
            return jsonify({'message': 'Token is invalid!'}), 401

        return f(user_id, *args, **kwargs)
    return decorated