requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
Werkzeug
psycopg2-binary
flask-cors